move derive_delta to crypto
